samus: Do not put touchscreen in reset based on lid switch
If we put the touchscreen in reset solely based on the lid state then
we do not allow the kernel driver to properly save and restore controller
state when entering suspend from a lid close event.

Since we want to always have touchscreen in reset in suspend/off state
and not in reset when the system is on then we can just remove this hook
because there is already logic in power_handle_state for that behavior.

BUG=chrome-os-partner:35879
BRANCH=samus
TEST=manual testing on samus:
1) put the device to sleep with lid open and powerd_dbus_suspend,
verify that TOUCHSCREEN_RESET_L=0
2) put the device to sleep with lid close and verify that
TOUCHSCREEN_RESET_L=0
3) check for kernel errors to ensure that there are no issues
saving touchscreen controller registers in the suspend path.

Change-Id: I5bff6356c54f064879a164e9ad4e67f1b2bcce6a
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>