WIP:samus: automatically recover from bad charging state
Detect and automatically recover from charging circuit not working.

Work in progress. DO NOT COMMIT. Need to make it robust enough to
make sure the recover doesn't kick in repeatedly and cause problems.
Also need to do something to make it such that the users override
setting is remembered and restored.

BUG=chrome-os-partner:35491
BRANCH=samus
TEST=use "charger voltage <mv>" where mv is a couple hundred mV less
than battery voltage to cause charging circuit to wedge, and then
wait and see that it is automaticlaly recovered.

Change-Id: I0b613ce67a077cbc82275961ebfe4b84a394ad12
Signed-off-by: Alec Berg <alecaberg@chromium.org>