Update patch set 1
Patch Set 1:

(2 comments)

I'm open to other suggestions on how to fix this. But basically sometimes we read some bad value for battery SOC, and that causes us to go into IDLE, which causes the charge circuit to get wedged and not allow charging again until AC is unplugged and replugged.

One other option is to turn on the PEC code, which I fiddled with briefly, but this seems like a more major change that needs a lot more testing.

Patch-set: 1