Update patch set 1
Patch Set 1: Code-Review+2

Given that there's a request to get this in tonight, I'll give it a +2.  I spent 30 minutes looking this over and it looks reasonable.  Please have Randall (or someone else on the EC team) give this a good review when it goes into master.  :)

As I understand it:
* Power button press comes in and EC and AP both respond
* AP quickly turns off.  That makes power good line go low.  
* EC responds to power good going low and does a power state transition.
* In the power state transition we get confused because the power button is still being pressed.

Thanks!

Patch-set: 1
Label: Code-Review=+2