Update patch set 4
Patch Set 4: Code-Review-1

It's messier than this.

Nuvoton uses external SPI flash, but directly maps it to the EC's address space so it appears internal.

So we need to think harder about the naming.  Because this change would have NPCX do:

#define CONFIG_FLASH_INTERNAL // but not actually internal
#undef CONFIG_FLASH_SPI // but it really is SPI flash

As a start, we could do:

CONFIG_FLASH_MAPPED --> flash is memory-mapped to EC's address space (that is, hashing code can directly read it)

That would be defined by default, but not for MEC.

I think rather than put your code in common/flash_spi.c, we should put that in chip/mec1322/flash.c.  And then we can refactor common parts up to common/ at a later point.

Patch-set: 4