Update patch set 1
Patch Set 1:

Im ok with this for samus and it solves a real problem. We need to add a note that more care will be needed on any future devices that have an OTG / device mode. In those cases the SDP that we detect could be a host that makes a USB connection to the OTG/device and this CL would interrupt that connection after 30s.

Patch-set: 1
Reviewer: Gerrit User 1001164 <1001164@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0