Update patch set 1
Patch Set 1: Code-Review+2

It wouldn't be that hard to change this to be handled in an interrupt. We already have the interrupt handler: i2c_event_handler(). You just need to enable RXIE, and massage the code a little to wait for the interrupt instead of using wait_isr().

Patch-set: 1
Label: Code-Review=+2