Update patch set 1
Patch Set 1: -Code-Review

I'm not sure if we should land this or not.

It does not fully fix my original problem(ideally I want to fix that with the next patch), maybe we should land it so it's a temporary fix, and i can also see it fixing other isr waits(ex. STM32_I2C_ISR_STOP), it should not cause problems after the next CL lands either.

Should we just land it?

Patch-set: 1