Update patch set 6
Patch Set 6: Code-Review+1 Verified+1

I found the cause of my hard faults.

Seems that even when CONFIG_TASK_PROFILING is defined, we still have a case near startup where need_resched_or_profiling could be 0 (this is only set to 1 when the tasks start). So we can't always trigger the svc in the irqs or we might trigger it when we didn't initialize the tasks yet.

Given that we still need an if statement(for at least a need_profiling) in the DECLARE_IRQ, I recommend leaving it as is.

Patch-set: 6
Label: Code-Review=+1
Label: Verified=+1