Minnie: Configure accelerometers
Add some config for minnie to support lid angle calculations.

The rot_hinge vectors are used to calculate the estimated vector if
the lid was open to 90 deg and 270 deg. This is later used to disambiguate
an equation when calculating the lid angle. Note that rot_hinge is wrong on
samus, the 2 cross products were always equal to each other, so there was no
disambiguation possible.

BUG=chrome-os-partner:37106
TEST=Have fun with accelread
TEST=Note that accelread 0 and accelread 1 show almost the same calibrated
value when the lid is closed (accelerometers aligned)
TEST=Throw in CPRINTF("Lid angle = %d\n", lpc_data[0]);
in common/motion_sense.c at line 163 and note how angle goes over 180 deg
TEST=(re: why it's different from samus) Add printouts of the estimated
vectors and compare them to the actual vector at 90 and 270 deg
BRANCH=firmware-veyron-6588.B

Change-Id: I4f6f9bf7446af30295492665a82d72ce263ba47f
Signed-off-by: james_chao <james_chao@asus.com>
Signed-off-by: Alexandru M Stan <amstan@chromium.org>