pd_log: Add command to request PD MCU to write a log
When we find that charging is in a wedged state, we may wish to write a
PD log entry, but the PD MCU cannot detect such a state on its own.
Therefore, add a new command to ask the PD MCU to write a log of a given
type.

BUG=chrome-os-partner:36668
TEST=Manual on samus:
./ectool --dev=1 pdwritelog charge 0
./ectool --dev=1 pdwritelog charge 1
./ectool --dev=1 pdwritelog connect 0
./ectool --dev=1 pdwritelog wedged 0
./ectool --dev=1 pdprintlogs
Verify log output matches expectation:
2015-02-12 10:09:48.136 P0 SRC
2015-02-12 10:09:48.142 P1 SNK Charger PD 20148mV max 20000mV / 3000mA
2015-02-12 10:09:48.150 P0 New connection
2015-02-12 10:09:48.157 P0 Wedged charge circuit detected
--- END OF LOG --
Also, verify kernel logging of wedged event:
[  121.168879] PDLOG 2015/02/12 18:11:45.452 P0 Event 02 (0000) []
BRANCH=Samus
Change-Id: I55c7c839cf8300fcd3931dccdaaf16c1065e31a8
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>