cortex-m*: Save panicinfo on div0 and stack overflow panics
Jump to exception_panic rather than to panic_reboot directly so that a
panic dump is logged. Also, log the panic type in r4, and the overflow
task (for stack overflow only) in r5 so that panic reasons can be
distinguished.

BUG=chrome-os-partner:36744
TEST=Manual on samus_pd. Run 'crash divzero' then 'panicinfo' after
reboot. Verify that panic info is printed with "r4 :dead6660". Trigger
stack overflow, verify that panic info is printed with "r4 :dead6661".
BRANCH=Samus

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Change-Id: I5f7a8eb0a5c2ac5799d29bb241deb24fabf38f68