Update patch set 1
Patch Set 1:

(2 comments)

Running average... Hm.

That's an interesting point. So, check every second, and just change the fans when the running average goes above the threshold. But then I'd worry about how to weight each input - if we have 6 board sensors, does that mean the PECI can get up to 100C before the average is high enough to turn the fans on? Or do we need individual limits PLUS the running average?

Patch-set: 1