Update patch set 3
Patch Set 3:

Okay, remember we know that the fans will keep things cool. The entire point of this CL is so that Andrew stops saying "When I'm in a completely quiet room, I can hear the fans turning off and on every few seconds."

If this is going to be bikeshedded into completely rewriting the algorithm, I'm going to let someone else implement it.

Patch-set: 3