Update patch set 2
Patch Set 2:

Excellent, I just put up a conflicting CL that should be ignored for now (I'll unmark it as ready) that was dealing with this for the USART code.

Speaking of which, this CL should also change the use of clock_get_freq in the chip/stm32/usart-stm32l.c code.

Thanks.

Patch-set: 2