Update patch set 3
Patch Set 3: Code-Review-2

I'm not sure. It's very useful to have, but very ugly. I'm not sure if it's worth having it, given that you need a disassembly(with debug symbols) for the output of this to be useful.

Patch-set: 3
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2