Veyron: Do not block power state machine by waiting for the power button release
The firmware needs to talk to the EC while the power button is pressed.
If the EC did not even leave the S5->S3 state this is not possible.

Seems like the piece of code is not even necessary, check_for_power_off_event
will catch the long press asynchronously later on anyway.

BUG=chrome-os-partner:36924
TEST="dut-control pwr_button:press sleep:2 pwr_button:release"
Still see "[32.888735 HC 0x17]" immediatelly after
"[32.502644 power state 3 = S0, in 0x0001]"
HC 0x87 should not be the first.
TEST="dut-control pwr_button:press sleep:9 pwr_button:release"
still works for it to shutdown, so that piece of code was totally useless.
BRANCH=veyron

Change-Id: I1e58b7cf6ce1fbf95d2128e8c9c11f4513da2cee
Signed-off-by: Alexandru M Stan <amstan@chromium.org>