Update patch set 1
Patch Set 1:

I tested both, but I couldn't tell whether this is better or worse than the previous version (CL 250501). It's different, but not by a lot, so it's likely to be a matter of personal preference.

I'm going to just leave this CL sitting here for a few weeks to see if the bug stays resolved.

Patch-set: 1