Update patch set 2
Patch Set 2:

(5 comments)

For the F411, I was using the clock to save power: we can not change the PLL easily, but we can go back to the HSI clock (16MHz).
For the sensor hub I think we need to be running in suspend mode. I can investigate if I can wake up the sensor hub on interrupt for the bosj sensor when it detects interesting pattern.

Using static is not trivial given the interface is common for all chips.

Patch-set: 2
Label: Code-Review=0