cr50: avoid shell errors when building for host
When building hosttests, CONFIG_FLASH_BASE is not a numeric
constant (it's a pointer to a buffer). This makes the the shell
complain when the Makefile tries to convert it from hex to
decimal.

Since the Makefile only needs that value to convert .bin to .hex,
let's just wait to calculate it until we need it.

BUG=chrome-os-partner:37071
BRANCH=none
TEST=manual

Before, "make hosttests" printed a bunch of this:

  sh: ((uintptr_t)__host_flash): missing `)' (error token is "__host_flash)")

Now it doesn't.

Change-Id: If408eb347b4f2385893d53cdfbf8fd5033868737
Signed-off-by: Bill Richardson <wfrichar@chromium.org>