DO NOT SUBMIT - npcx romlib
This hacks the build to create ROM-able functions and data at the
start of ec.RO.flat.  There's probably a more elegant way of doing
this, but this is sufficient for the current build.

The first 0x2000 bytes of ec.RO.flat can be placed in ROM at address
0x80.

The rest of ec.RO.flat is in the wrong place to run.  That's ok,
because the whole point of this hack is to get the first 8KB where I
want it.

BUG=none
BRANCH=none
TEST=make BOARD=npcx_evb
     Confirm functions and data in the first 8KB of the image are where they
     should be by inspecting ec.RO.flat, ec.RO.map, ec.RO.dis

Change-Id: I7eda921c5e4676982cdce4c00372c3fbc2a803ad
Signed-off-by: Randall Spangler <rspangler@chromium.org>