Update patch set 9
Patch Set 9:

How critical is this change? Do we need this for the updater or is this just cleanup?

I don't want to land a bunch of new stuff that has not been well tested for the updater we are trying to get going in the near future.

Patch-set: 9
Reviewer: Gerrit User 1000868 <1000868@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0