Update patch set 1
Patch Set 1: Code-Review+2

Thanks for finding this, I really screwed it up. EC_CMD_USB_PD_SET_AMODE is sent only from ectool, so we're safe there. EC_CMD_PD_WRITE_LOG_ENTRY is sent by the EC to the PD when charging is wedged. Once this lands in RW, if the EC is in RW and PD is in RO, the command will be misinterpreted as a EC_CMD_USB_PD_SET_AMODE command. I'm not sure if this can cause harm, since the param structures are different I'll have to overlay them and see what case we can fall into. We may need to bump the version of EC_CMD_PD_WRITE_LOG_ENTRY from the sender and receiver.

Patch-set: 1