Minnie: Define rot_hinge_{90,180} transforms
The rot_hinge vectors are used to calculate the estimated vector if
the lid was open to 90 deg and 270 deg. This is later used to disambiguate
an equation when calculating the lid angle.

Since these vectors were wrong, the 2 cross products were always equal to
each other, so there was no disambiguation to be done.

BUG=chrome-os-partner:37106
TEST=Throw in CPRINTF("Lid angle = %d\n", lpc_data[0]);
in common/motion_sense.c at line 163 and note how angle goes over 180 deg
TEST=Add printouts of the estimated vectors and compare them to the actual
vector at 90 and 270 deg
BRANCH=veyron

Change-Id: I662b114dbaafb4d66419bb146c5e08ee3899d458
Signed-off-by: Alexandru M Stan <amstan@chromium.org>