Update patch set 1
Patch Set 1:

Answering for Kevin "Can you explain (3)? "The reason we pick 256KB instead of physcial SPI Size was to increase the flash update speed on EC SW Sync and OS Flashrom.""
If we had a prototype board that had a large capacity flash memory attached to an EC, then programming all of it either at boot time (via software sync) or via other means such as the servo would take a long time. If we used facilities such as the flash map support in flashrom then this hypothetical prototype board would not encourage us to throw it out of the window every time we wanted to update the EC image. Note, we typically do not use SPI flash emulators here for some reason.

Patch-set: 1