Update patch set 2
Patch Set 2:

Renamed from flash_private_spi_pp3300 to flash_mec1322 since the file padding with image at top of SPI is specific to MEC1322. Other EC could place the image at any location of the SPI chip.

Removed save/restore servo settings since we do not know what is the previous spi1_vref value, so flash_mec1322 will force spi1_vref to off to ensure the servo is not driving any voltage to the board after programming is done. And flash_mec1322 will force SPI1 interface to OFF so that the board will be in a bootable state.

Patch-set: 2