mighty: fix the LED bug of removed battery
Clear curr.batt_is_charging when the battery is not present.

Add a new batt_was_removed flag to track when the battery has been
removed.  This causes PWR_STATE_ERROR so the LED shows error state,
and triggers re-reading the static parameters from the battery when
it's reattached.

BUG=chrome-os-partner:38235
TEST=check the LED state on mighty
BRANCH=veyron

Change-Id: I46a7a9ccbe1bd2c10cdeeb565632b90fe9c4cc46
Signed-off-by: ZhengShunQian <zhengsq@rock-chips.com>