Update patch set 3
Patch Set 3:

Tried, and the kernel still crashes.
The issue is the ioctl handler has a buffer set to PROTO v2 size and no check are done on insize when calling cmd_xfer.
Therefore if the caller and the EC agree on a big size, the lower layer will corrupt memory...

The answer is to make the ioctl handler smarter, have a small size buffer ready for 90% of the call and the ability to allocate a bigger buffer if necessary.

Patch-set: 3