Update patch set 3
Patch Set 3:

> (1 comment)

yes, lfw links with .o-s that are compiled against main board gpio_list. Which can be huge.

So the idea here is to shrink LFW's GPIOs to bare minimum. Then real board's gpio.inc need to be updated in such a way that gpio definitons needed for LFW come first. That way LFW gets away with just a few GPIOs AND can link against .o-s.

I agree it is not pretty to have this sort of limitation. But alternative would be to recompile some .o-s and add ifdefs in gpio.inc. Definitely doable, but we are getting into adding complexity info makefiles. To me keeping gpio.inc in sync sounded like less evil.

I am open to any ideas, if 2-stage compilation is better lets redo this bit.

Patch-set: 3