Update patch set 4
Patch Set 4:

> Hmm.  If you're compiling some module only once, with the main
 > firmware's GPIO enum list, and linking it against both main and
 > lfw, then won't the shrunk list from lfw access the wrong GPIO
 > table entries?
 
I see no problem here. gpio_list is accessed by index. As long as indices are the same both lfw and main code can be linked together. Since we want to save space in lfw's gpio_list, we move all the needed entries in front of the table. That way we have both the same indices and shrunk space.

gpio_list entries do not reference each other. So I can't think of a problem with the approach suggested. Am I missing something? 
 
 > But it's also not great to force all the lfw GPIOs first.  We
 > currently optimize GPIO interrupt handling by putting all the GPIOs
 > with interrupts first.
 
Can't argue here. We would have to sacrifice this optimization to an extent of (currently) 1 gpio used by lfw. Since the number of gpios is likely to be small, wouldn't be acceptable? We probably waste 10 or so instructions per skipped gpio.

The change suggested here tries pretty hard to avoid 2-stage compilation and complexity associated with it. If we want gpio definition to be clean and quirks-free, we are buying ourselves into more complex makefiles. Lets decide what is preferable, or is there a cleaner solution?

Patch-set: 4