Update patch set 2
Patch Set 2:

I don't suppose checkpatch has a comment syntax to toggle warnings around a block of code.  Then we could leave the warning enabled, just not in the USB code.

Are the structs copy-pasted from somewhere else?  I'm guessing they are.  They're not very long structs, though.  Perhaps it would be better to rename the members to kernel style so coding style is consistent.

Patch-set: 2