Update patch set 2
Patch Set 2:

> I don't suppose checkpatch has a comment syntax to toggle warnings
 > around a block of code.  Then we could leave the warning enabled,
 > just not in the USB code.

Not that I've been able to find, and it doesn't quite sound like something that Linus would be OK with.  :)

 > Are the structs copy-pasted from somewhere else?  I'm guessing they
 > are.  They're not very long structs, though.  Perhaps it would be
 > better to rename the members to kernel style so coding style is
 > consistent.

They are named this way to conform with the naming of the fields in the USB descriptors and ALL USB documentation.  So I'm not inclined to change them.  Since this is really just an issue when working with USB descriptors I'd be fine not committing this change, and living with the extra spew now and then...

Patch-set: 2